requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
moviepy==2.1.2
openai-whisper==20240930
langchain==0.3.19
langchain-postgres==0.0.13
langchain-cohere==0.4.2
psycopg2-binary==2.9.10
cohere==5.14.0
sqlalchemy==2.0.38
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.162.0
python-dotenv==1.0.1
pyngrok==7.2.3
torch==2.6.0
python-jose==3.4.0
passlib==1.7.4
pydantic==2.10.6
bcrypt==4.3.0
pydantic[email]
pytube==15.0.0